rename project to 'com.androidimpact.app'
