+++ app/src/androidTest/java/com/androidimpact/app/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.androidimpact;
+package com.androidimpact.app;
 
 import android.content.Context;
