+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.AndroidImpact"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.androidimpact.app.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
